requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.0.0
transformers>=4.30.0
PyPDF2>=3.0.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
groq>=0.4.0
langchain-groq>=0.1.0
tqdm>=4.66.0
faiss-cpu>=1.7.4
langchain-ollama>=0.1.0
streamlit-option-menu>=0.3.0
chromadb>=0.4.24
